Try to do base 7 homework
