+++ buttle.js
@@ -1,0 +1,40 @@
+import { getRandom } from "./utiits.js";
+
+const $formFight = document.querySelector(".control");
+
+export const HIT = {
+  head: 30,
+  body: 25,
+  foot: 20,
+};
+
+export const ATTACK = ["head", "body", "foot"];
+
+export const enemyAttack = () => {
+  const hit = ATTACK[getRandom(3) - 1];
+  const defence = ATTACK[getRandom(3) - 1];
+
+  return {
+    value: getRandom(HIT[hit]),
+    hit,
+    defence,
+  };
+};
+
+export const playerAttack = () => {
+  const attack = {};
+
+  for (let item of $formFight) {
+    if (item.checked && item.name === "hit") {
+      attack.value = getRandom(HIT[item.value]);
+      attack.hit = item.value;
+    }
+    if (item.checked && item.name === "defence") {
+      attack.defence = item.value;
+    }
+
+    item.checked = false;
+  }
+
+  return attack;
+};
