+++ logs.js
@@ -1,0 +1,87 @@
+import { getRandom } from "./utiits.js";
+
+const $chat = document.querySelector(".chat");
+
+export const logs = {
+  start:
+    "Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.",
+  end: [
+    "Результат удара [playerWins]: [playerLose] - труп",
+    "[playerLose] погиб от удара бойца [playerWins]",
+    "Результат боя: [playerLose] - жертва, [playerWins] - убийца",
+  ],
+  hit: [
+    "[playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага.",
+    "[playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника.",
+    "[playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента.",
+    "[playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента.",
+    "[playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента.",
+    "[playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага.",
+    "[playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника.",
+    "[playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника.",
+    "[playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника.",
+    "[playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника.",
+    "[playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника.",
+    "[playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага.",
+    "[playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента.",
+    "[playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника.",
+    "[playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента.",
+    "[playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента.",
+    "[playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника.",
+    "[playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага.",
+  ],
+  defence: [
+    "[playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.",
+    "[playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.",
+    "[playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.",
+    "[playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.",
+    "[playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.",
+    "[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.",
+    "[playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.",
+    "[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.",
+  ],
+  draw: "Ничья - это тоже победа!",
+};
+
+export function generateLogs(type, who_hit, who_defence, damage) {
+  const date = new Date();
+  const time = date.toLocaleTimeString();
+  switch (type) {
+    case "start":
+      const start = logs[type]
+        .replace("[player1]", who_hit.name)
+        .replace("[player2]", who_defence.name)
+        .replace("[time]", time);
+      const elStart = `<p>${start}</p>`;
+      $chat.insertAdjacentHTML("afterbegin", elStart);
+      break;
+    case "hit":
+      const hit = logs[type][getRandom(17)]
+        .replace("[playerKick]", who_hit.name)
+        .replace("[playerDefence]", who_defence.name);
+      const elHit = `<p>${time} ${hit} -${damage} ${who_defence.hp}/100</p>`;
+      $chat.insertAdjacentHTML("afterbegin", elHit);
+      break;
+    case "defence":
+      const defence = logs[type][getRandom(7)]
+        .replace("[playerKick]", who_hit.name)
+        .replace("[playerDefence]", who_defence.name);
+      const elDefence = `<p>${time} ${defence}</p>`;
+      $chat.insertAdjacentHTML("afterbegin", elDefence);
+      break;
+    case "end":
+      const end = logs[type][getRandom(2)]
+        .replace("[playerWins]", who_hit.name)
+        .replace("[playerLose]", who_defence.name);
+      const elEnd = `<p>${time} ${end}</p>`;
+      $chat.insertAdjacentHTML("afterbegin", elEnd);
+      break;
+    case "draw":
+      const draw = logs[type][0];
+      const elDraw = `<p>${time} ${draw}</p>`;
+      $chat.insertAdjacentHTML("afterbegin", elDraw);
+      break;
+    default:
+      break;
+  }
+}
