+++ utiits.js
@@ -1,0 +1,4 @@
+export const getRandom = (num) => {
+  return Math.ceil(Math.random() * num);
+};
+
